requirements.txt
fastapi
uvicorn
torch
transformers
pillow
easyocr
pytesseract
scipy
huggingface-hub
python-multipart
sacremoses
protobuf
